Fixed possible crash on death

+++ src/game/server/entities/character.cpp
@@ -693,7 +693,7 @@ void CCharacter::Die(int Killer, int Weapon)
 			Killer, - 1 - Killer,
 			m_pPlayer->GetCID(), m_pPlayer->GetTeam(), Server()->ClientName(m_pPlayer->GetCID()), Weapon, ModeSpecial
 		);
-	else
+	else if(GameServer()->m_apPlayers[Killer])
 		str_format(aBuf, sizeof(aBuf), "kill killer='%d:%d:%s' victim='%d:%d:%s' weapon=%d special=%d",
 			Killer, GameServer()->m_apPlayers[Killer]->GetTeam(), Server()->ClientName(Killer),
 			m_pPlayer->GetCID(), m_pPlayer->GetTeam(), Server()->ClientName(m_pPlayer->GetCID()), Weapon, ModeSpecial
